Fix crash when reloading config.rb
The previous implementation cached the reactor on the extension instance. However, the instance doesn’t survive the reload and would try to create a new reactor. This failed, because the previous reactor wasn’t stopped.

This change caches the reactor in the ractor class itself so it’s still there after the reload.